Final version

+++ ft_atoi.c
@@ -6,28 +6,27 @@
 /*   By: psitkin <psitkin@hive.student.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/10 23:51:07 by psitkin           #+#    #+#             */
-/*   Updated: 2023/11/17 00:51:18 by psitkin          ###   ########.fr       */
+/*   Updated: 2023/11/19 19:30:00 by psitkin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-#include <stdio.h>
 
-static int	convert_nb(const char *str, long nb, int sign)
+static int	convert_nb(const char *str, size_t nb, int sign)
 {
 	int	i;
 
 	i = 0;
 	while (str[i] && (str[i] >= 48 && str[i] <= 57))
 	{
-		if (nb > LONG_MAX / 10)
+		if (nb > SIZE_MAX / 10)
 		{
 			if (sign < 0)
 				return (0);
 			return (-1);
 		}
 		nb = nb * 10;
-		if (nb > LONG_MAX - (str[i] - 48))
+		if (nb > SIZE_MAX - (str[i] - 48))
 		{
 			if (sign < 0)
 				return (0);
@@ -41,9 +40,9 @@ static int	convert_nb(const char *str, long nb, int sign)
 
 int	ft_atoi(const char *str)
 {
-	size_t	i;
-	long	nb;
-	int		sign;
+	size_t		i;
+	size_t		nb;
+	int			sign;
 
 	nb = 0;
 	i = 0;
